New branch for virtualized os package (vos) This package addresses two major limitations in the Layercake's pretend-mode implementation and in the unit-testing framework. The existing pretend mode is often inaccurate because it does not have a lot of knowledge of existing mounts and open files whereas the unit-testing framework fails to account for the interactions of mounts and files.  The current vos package is the result of several iterations of a design to address these problems.
Client code within Layercake, Stagemaker, and the units tests is
to be set to refer to an instance of a vos-object receiver rather
than the os package of the standard library.  The vos variable may
be set to an object that proxies the native os services or to a
virtual-OS object (type *MemOS) which manages the structures in
memory.  The MemOS object may be set so that reads are directed to
the native file system; this mode of operation is to support
Layercake's shadowing of the real filesystem in pretend mode.

Working now:  in-memory inode and storage-oriented filesystem
operations including all file and directory operations required by
Layercake and Stagemaker, devtmpfs creation and mounts, process
creation and teardown, and unit tests to exercise all of these.
Not yet implemented:  procfs, bind and overlay mounts, and reads
from the shadowed filesystem.

Important:  MemOS does no locking and so a MemOS instance is not
expected to tolerate simulaneous access from multiple goroutines.
No such locking capability is contemplated.  Since MemOS does not
use global state, simultaneous operation on separate MemOS objects
is OK.